support additional instructions for waitloop skipping
Adds support for MultipleLoadStore, LoadStoreImmediateOffset, and
MoveShiftedRegister. This allows waitloop skipping to work in Spiderman
the Movie, bringing in-game fps from 280 to 375 and title screen fps
from 450 to 700.

This change also disallows instructions that branch from participating
in waitloop skipping. I don't expect this to affect anything, but I
figure it's better to be overly cautious.